% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ctmcViterbi}
\alias{ctmcViterbi}
\title{Viterbi algorithm implementated in Rcpp}
\usage{
ctmcViterbi(ts, theta, obs)
}
\arguments{
\item{ts}{size m NumericVector of timestamp associated with each observed outcome}

\item{theta}{a double type parameter for the calculation of transition probability}

\item{obs}{a NumericMatrix with # of columns the same as the length of ts, and # of rows the same as # of possible states, each element is the probability for observed outcome given each state at each timestamp}
}
\value{
a vector of length m (the length of ts), representing the Viterbi path of maximum likelihood
}
\description{
This function implements Viterbi algorithm on a continuous-time	HMM, given a vector of timestamps, parameter theta for transition probability function, and a matrix of probability for observed outcome given each state at each timestamp
}
\examples{
obs <- matrix(c(0.88,0.10,0.88,0.10,0.02,0.30,0.02,0.30,0.10,0.60),2,5)
theta <- log(2)
ts <- c(1,2,3,4,5)
ctmcViterbi(ts,theta,obs)
}
\references{
BIOS 615 lecture notes
}
