% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jinyueHW4-package.r
\docType{package}
\name{jinyueHW4}
\alias{jinyueHW4}
\alias{jinyueHW4-package}
\title{jinyueHW4.}
\description{
An implementation of two HMM algorithms in Rcpp
}
\details{
The package includes two functions for continuous time HMM: a Viterbi algorithm implementation and a forward-backward algorithm implementation
}
\author{
Yue JIN
}
